init dir

+++ install.sh
@@ -9,7 +9,12 @@ git config --global alias.last 'log -1 HEAD'
 git config --global alias.ll log --oneline
 git config --global alias.d diff
 
-lsd_deb_name="lsd_0.21.0_amd64.deb"
-lsd_repo_url="https://github.com/Peltoche/lsd"
-wget "$lsd_repo_url/releases/download/0.21.0/$lsd_deb_name" && dpkg -i $lsd_deb_name && rm $lsd_deb_name
+if ! command -v lsd &> /dev/null
+then
+    lsd_deb_name="lsd_0.21.0_amd64.deb"
+    lsd_repo_url="https://github.com/Peltoche/lsd"
+    wget "$lsd_repo_url/releases/download/0.21.0/$lsd_deb_name" && dpkg -i $lsd_deb_name && rm $lsd_deb_name
+fi
 
+cat "$PWD/init/.bash_history" >> ~/.bash_history
+cat "$PWD/init/.config/shortcut-locations" >> ~/.config/shortcut-locations
